+++ fili-core/src/main/java/com/yahoo/bard/webservice/application/AbstractBinderFactory.java
@@ -178,6 +178,7 @@ public abstract class AbstractBinderFactory implements BinderFactory {
 
     private final TaskScheduler loaderScheduler = new TaskScheduler(LOADER_SCHEDULER_THREAD_POOL_SIZE);
 
+
     /**
      * Constructor.
      */
@@ -230,6 +231,9 @@ public abstract class AbstractBinderFactory implements BinderFactory {
 
                 //Initialize the metrics filter helper
                 FieldConverterSupplier.metricsFilterSetBuilder = initializeMetricsFilterSetBuilder();
+
+                // Build the datasource metadata service containing the data segments
+                bind(getDataSourceMetadataService()).to(DataSourceMetadataService.class);
 
                 // Build the datasource metadata service containing the data segments
                 bind(getDataSourceMetadataService()).to(DataSourceMetadataService.class);